Make handleFindContent abstract in baseProtocol
